new architecture for DWI pipelines
