+++ clinica/pipelines/dwi/__init__.py
@@ -1,0 +1,1 @@
+from . import connectome, dti, preprocessing
