+++ clinica/pipelines/dwi/connectome/info.json
