+++ test/unittests/utils/test_input_files.py
@@ -1,9 +1,8 @@
-import os
 from pathlib import Path
 
 import pytest
 
-from clinica.utils.dwi import DTIBasedMeasure
+from clinica.pipelines.dwi.dti.utils import DTIBasedMeasure
 from clinica.utils.pet import ReconstructionMethod, Tracer
